Move type definitions to entrypoints/
Only TS 4.7 will resolve them now, but it's cleaner.
