--- entrypoints/index.d.cts
@@ -1,2 +1,0 @@
-export {default} from '../index.js';
-export * from '../index.js';
