+++ entrypoints/main.d.cts
@@ -1,0 +1,2 @@
+export {default} from './main.js';
+export * from './main.js';
