+++ entrypoints/plugin.d.cts
@@ -1,1 +1,1 @@
-export * from '../plugin.js'
+export * from './plugin.js'
