+++ entrypoints/plugin.d.ts
